Correctly match URLs with no stats type
Make the stats type optional. The original reason for not catching that
case is that /stats/example.wordpress.com would redirect to
/stats/insights/example.wordpress.com, but this appears to be a
JavaScript redirect and would not get caught by the script.
Additionally, since match returns null when there is no match, treating
parsedUrl as an array to extract the stats type and the blog domain will
fail.